Move OrderItemQuantityField into OrderItem namespace
